Refactor MathML visitors to make MathMLVisitor work like SerializedMmlVisitor, and to move common code to the base MmlVisitor class.  (mathjax/MathJax#2920)
